Add files via upload
Contains two folders. Jenkins and scripts. Jenkins contain the Dockerfile and start up script to set up the Jenkins environment. Scripts contains three folders. One of the folders, EKS, have the start and stop scripts for EKS. Another folder is NGINX, which has the Dockerfile and Kubectrl scripts to deploy a simple NGINX app on EKS. While the last is Streamlit, is similar to the NGINX folder. It too contains a Dockerfile and Kubectrl scripts to deploy the app on EKS.